Updated the get()
Added an if statement to check if the entered index is less 0 or greater than the size of the array. This would create an IndexOutOfBoundsException. Also instead of printing null to the concept, I add a null check with this condition. While the indices that are not filled will be null, the indices outside of the range should not return null.